Gui: slurm users can now be listed
