+++ usermgmt_gui/src/current_selected_view/listing_state.rs
@@ -7,4 +7,5 @@ pub struct ListingState {
     pub rw_user_name: Option<String>,
     pub rw_pw: Option<String>,
     pub list_ldap_res: IoResourceManager<LdapSearchResult>,
+    pub list_slurm_user_res: IoResourceManager<String>,
 }
